refactor: split task schedule service modules (#4897)
