+++ server/services/taskScheduleTemplates.js
@@ -1,0 +1,50 @@
+/** User-defined task templates stored with the task schedule. */
+
+import { emitLog } from './cosEvents.js';
+import { loadSchedule, saveSchedule } from './taskScheduleStore.js';
+
+// ============================================================
+// Templates
+// ============================================================
+
+export async function addTemplateTask(template) {
+  const schedule = await loadSchedule();
+
+  const newTemplate = {
+    id: `template-${Date.now().toString(36)}`,
+    createdAt: new Date().toISOString(),
+    name: template.name,
+    description: template.description,
+    category: template.category || 'custom',
+    taskType: template.taskType,
+    priority: template.priority || 'MEDIUM',
+    metadata: template.metadata || {}
+  };
+
+  schedule.templates.push(newTemplate);
+  await saveSchedule(schedule);
+
+  emitLog('info', `Added template task: ${newTemplate.name}`, { templateId: newTemplate.id }, '📅 TaskSchedule');
+  return newTemplate;
+}
+
+export async function getTemplateTasks() {
+  const schedule = await loadSchedule();
+  return schedule.templates;
+}
+
+export async function deleteTemplateTask(templateId) {
+  const schedule = await loadSchedule();
+  const index = schedule.templates.findIndex(t => t.id === templateId);
+
+  if (index === -1) {
+    return { error: 'Template not found' };
+  }
+
+  const deleted = schedule.templates.splice(index, 1)[0];
+  await saveSchedule(schedule);
+
+  emitLog('info', `Deleted template task: ${deleted.name}`, { templateId }, '📅 TaskSchedule');
+  return { success: true, deleted };
+}
+
